Add auth-aware size caps fetch and send logs
